+++ system/ping6/ping6.c
@@ -53,8 +53,8 @@
 struct ping6_priv_s
 {
   int code;                        /* Notice code ICMP_I/E/W_XXX */
-  int tmin;                        /* Minimum round trip time */
-  int tmax;                        /* Maximum round trip time */
+  long tmin;                       /* Minimum round trip time */
+  long tmax;                       /* Maximum round trip time */
   long long tsum;                  /* Sum of all times, for doing average */
   long long tsum2;                 /* Sum2 is the sum of the squares of sum ,for doing mean deviation */
 };
@@ -128,7 +128,7 @@ static void ping6_result(FAR const struct ping6_result_s *result)
         break;
 
       case ICMPv6_E_SOCKET:
-        fprintf(stderr, "ERROR: socket() failed: %d\n", result->extra);
+        fprintf(stderr, "ERROR: socket() failed: %ld\n", result->extra);
         break;
 
       case ICMPv6_I_BEGIN:
@@ -139,44 +139,44 @@ static void ping6_result(FAR const struct ping6_result_s *result)
         break;
 
       case ICMPv6_E_SENDTO:
-        fprintf(stderr, "ERROR: sendto failed at seqno %u: %d\n",
+        fprintf(stderr, "ERROR: sendto failed at seqno %u: %ld\n",
                 result->seqno, result->extra);
         break;
 
       case ICMPv6_E_SENDSMALL:
-        fprintf(stderr, "ERROR: sendto returned %d, expected %u\n",
+        fprintf(stderr, "ERROR: sendto returned %ld, expected %u\n",
                 result->extra, result->outsize);
         break;
 
       case ICMPv6_E_POLL:
-        fprintf(stderr, "ERROR: poll failed: %d\n", result->extra);
+        fprintf(stderr, "ERROR: poll failed: %ld\n", result->extra);
         break;
 
       case ICMPv6_W_TIMEOUT:
         inet_ntop(AF_INET6, result->dest.s6_addr16, strbuffer,
                   INET6_ADDRSTRLEN);
-        printf("No response from %s: icmp_seq=%u time=%d ms\n",
+        printf("No response from %s: icmp_seq=%u time=%ld ms\n",
                strbuffer, result->seqno, result->extra);
         break;
 
       case ICMPv6_E_RECVFROM:
-        fprintf(stderr, "ERROR: recvfrom failed: %d\n", result->extra);
+        fprintf(stderr, "ERROR: recvfrom failed: %ld\n", result->extra);
         break;
 
       case ICMPv6_E_RECVSMALL:
-        fprintf(stderr, "ERROR: short ICMP packet: %d\n", result->extra);
+        fprintf(stderr, "ERROR: short ICMP packet: %ld\n", result->extra);
         break;
 
       case ICMPv6_W_IDDIFF:
         fprintf(stderr,
-                "WARNING: Ignoring ICMP reply with ID %d.  "
+                "WARNING: Ignoring ICMP reply with ID %ld.  "
                 "Expected %u\n",
                 result->extra, result->id);
         break;
 
       case ICMPv6_W_SEQNOBIG:
         fprintf(stderr,
-                "WARNING: Ignoring ICMP reply to sequence %d.  "
+                "WARNING: Ignoring ICMP reply to sequence %ld.  "
                 "Expected <= %u\n",
                 result->extra, result->seqno);
         break;
@@ -200,7 +200,7 @@ static void ping6_result(FAR const struct ping6_result_s *result)
 
         inet_ntop(AF_INET6, result->dest.s6_addr16, strbuffer,
                   INET6_ADDRSTRLEN);
-        printf("%u bytes from %s icmp_seq=%u time=%u.%d ms\n",
+        printf("%u bytes from %s icmp_seq=%u time=%ld.%ld ms\n",
                result->info->datalen, strbuffer, result->seqno,
                result->extra / USEC_PER_MSEC,
                result->extra % USEC_PER_MSEC / MSEC_PER_DSEC);
@@ -209,7 +209,7 @@ static void ping6_result(FAR const struct ping6_result_s *result)
       case ICMPv6_W_RECVBIG:
         fprintf(stderr,
                 "WARNING: Ignoring ICMP reply with different payload "
-                "size: %d vs %u\n",
+                "size: %ld vs %u\n",
                 result->extra, result->outsize);
         break;
 
@@ -218,7 +218,7 @@ static void ping6_result(FAR const struct ping6_result_s *result)
         break;
 
       case ICMPv6_W_TYPE:
-        fprintf(stderr, "WARNING: ICMP packet with unknown type: %d\n",
+        fprintf(stderr, "WARNING: ICMP packet with unknown type: %ld\n",
                 result->extra);
         break;
 
@@ -234,17 +234,17 @@ static void ping6_result(FAR const struct ping6_result_s *result)
                    result->nrequests;
 
             printf("%u packets transmitted, %u received, %u%% packet loss,"
-                   "time %d ms\n",
+                   "time %ld ms\n",
                    result->nrequests, result->nreplies, tmp,
                    result->extra / USEC_PER_MSEC);
             if (result->nreplies > 0)
               {
-                int avg = priv->tsum / result->nreplies;
-                int tmdev = sqrt(priv->tsum2 / result->nreplies -
+                long avg = priv->tsum / result->nreplies;
+                long tmdev = sqrt(priv->tsum2 / result->nreplies -
                                  (long long)avg * avg);
 
-                printf("rtt min/avg/max/mdev = %d.%03d/%u.%03u/"
-                       "%d.%03d/%d.%03d ms\n",
+                printf("rtt min/avg/max/mdev = %ld.%03ld/%ld.%03ld/"
+                       "%ld.%03ld/%ld.%03ld ms\n",
                        priv->tmin / USEC_PER_MSEC,
                        priv->tmin % USEC_PER_MSEC,
                        avg / USEC_PER_MSEC, avg % USEC_PER_MSEC,
